feat: inserido o atributo capacidade do drone em kg(Float), nas Classes Delivery e DeliveryDto, sendo implementadas em DeliveryService. E refatorado o atributo peso de entrega para float, nas classes Delivery e DeliveryDto.

+++ src/main/java/com/futureh/drone/feeder/dto/DeliveryDto.java
@@ -9,7 +9,7 @@ public class DeliveryDto {
   private String zipCode;
   private String longitude;
   private String latitude;
-  private Integer weightInKg;
+  private Float weightInKg;
 
   public String getAddress() {
     return address;
@@ -43,11 +43,11 @@ public class DeliveryDto {
     this.latitude = latitude;
   }
 
-  public Integer getWeightInKg() {
+  public Float getWeightInKg() {
     return weightInKg;
   }
 
-  public void setWeightInKg(Integer weightInKg) {
+  public void setWeightInKg(Float weightInKg) {
     this.weightInKg = weightInKg;
   }
