+++ src/main/java/com/futureh/drone/feeder/service/DeliveryService.java
@@ -21,7 +21,7 @@ public class DeliveryService {
     String zipCode = delivery.getZipCode();
     String longitude = delivery.getLongitude();
     String latitude = delivery.getLatitude();
-    Integer weightInKg = delivery.getWeightInKg();
+    Float weightInKg = delivery.getWeightInKg();
     Delivery newDelivery = deliveryRepository.save(
         new Delivery(address, zipCode, longitude, latitude, weightInKg));
     return newDelivery;
